bakhabazar/requirements.txt
django == 5.2.5
djangorestframework
hashids
Pillow
psycopg2-binary
gunicorn ==22.0.0
python-dotenv == 1.0.1
adrf
daphne
uvicorn[standard]
asyncio
aiohttp
openai